fix: relocate audit classes to engine.core after main merge
Move RiskEvaluationAuditConfig and RiskEvaluationAuditPublisher into
engine.core to align with upstream #70 and restore package access to
AbstractRiskEngine.EvaluatorResult.
