+++ core/src/test/java/io/github/mabartos/engine/core/RiskEvaluationAuditPublisherTest.java
@@ -1,4 +1,4 @@
-package io.github.mabartos.engine;
+package io.github.mabartos.engine.core;
 
 import io.github.mabartos.spi.level.Risk;
 import io.github.mabartos.spi.level.RiskLevel;
